requirements.txt
pymongo[srv]
mongoengine
python-decouple
xmltodict
black
requests
pyqt5
pyscrypt
